Module weedle::types

Structs

AttributedNonAnyType

Parses [attributes]? type where the type is a single non-any type

AttributedType

Parses [attributes]? type

DoubleType

Parses unrestricted? double

FloatType

Parses unrestricted? float

FrozenArrayType

Parses FrozenArray<Type>

LongLongType

Parses unsigned? long long

LongType

Parses unsigned? long

MayBeNull

Parses a nullable type. Ex: object | object??

PromiseType

Parses a Promise<Type|undefined> type

RecordType

Parses record<StringType, Type>

SequenceType

Parses sequence<Type>

ShortType

Parses unsigned? short

Enums

ConstType

Parses a const type

FloatingPointType

Parses unrestricted? float|double

IntegerType

Parses unsigned? short|long|(long long)

NonAnyType

RecordKeyType

Parses one of the string types ByteString|DOMString|USVString or any other type.

ReturnType

Parses the return type which may be undefined or any given Type

SingleType

Parses one of the single types

Type

Parses either single type or a union type

UnionMemberType

Parses one of the member of a union type

Type Definitions

UnionType

Parses a union of types